+++ src/main/java/ninja/eivind/hotsreplayuploader/services/AccountService.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package ninja.eivind.hotsreplayuploader;
+package ninja.eivind.hotsreplayuploader.services;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import javafx.concurrent.ScheduledService;
